Update science.md
Thank you for all of your comments. I've tried to broaden the scope by adding paragraphs on column water vapor and self aggregation; EarthCare, statistical sampling and relating process observations to climate change; and different ways of understanding the ITCZ (air-sea interactions, conservation of momentum/evolution of vorticity and divergence, etc.). Let me know if I've hit the mark or if some amendments are necessary.

If there is the energy and appetite for it, I think some of these paragraphs could be expanded on in separate posts. For instance, Divya and Marius could write about conservation of momentum/evolution of vorticity and divergence and the ITCZ; Jakob and I about cloud radiative effects; Helene and Nina about results from the dropsondes, etc.

By the way, who are the two people pictured?